feat: create new tables in the database

+++ src/main/java/ru/t1/java/demo/dto/TransactionDto.java
@@ -20,7 +20,7 @@ import java.sql.Timestamp;
 public class TransactionDto implements Serializable {
 
     @JsonProperty("accountId")
-    BigDecimal accountId;
+    Long accountId;
 
     @JsonProperty("amount")
     BigDecimal amount;
